pyproject.toml
[project]
name = "mental-sync-cli"
version = "0.1.0"
description = "Mental-Sync-CLI (MSC): A local-first, secure, and self-bootstrapping Agent Runtime Environment."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.10.0",
    "typer>=0.15.0",
    "rich>=13.9.0",
    "loguru>=0.7.0",
    "openai>=2.16.0",
    "pyyaml>=6.0.3",
    "anthropic>=0.77.1",
    "google-genai>=1.61.0",
    "httpx>=0.28.1",
]

[tool.uv.workspace]
members = ["packages/*"]

[tool.hatch.build.targets.wheel]
packages = ["src/msc"]

[project.scripts]
msc = "msc.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "RET", "SIM", "TCH", "ARG", "PTH", "ERA", "PL", "RUF"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
